@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  baseURL: 'https://twitter-python-backend-yc4c5cvnnq-df.a.run.app',
+  baseURL: 'http://api.ng-arab.tech/api/v1',
 };
