@@ -4,7 +4,7 @@
 
 export const environment = {
   production: false,
-  baseURL: 'https://twitter-python-backend-yc4c5cvnnq-df.a.run.app',
+  baseURL: 'http://api.ng-arab.tech/api/v1',
 };
 
 /*
